chore: relocate getfiletooltip
