+++ src/templates/workspace/snippets/listItemButtons.ts
@@ -1,6 +1,6 @@
 import { t } from 'vscode-ext-localisation';
-import { getFileTooltip } from '../../../utils/strings/getFileTooltip';
 import { RenderVars } from '../../../webviews/webviews.interface';
+import { getFileTooltip } from '../../../webviews/Workspace/helpers/getFileTooltip';
 import { File } from '../../../webviews/Workspace/WorkspaceViewProvider.interface';
 import { getImgUrls } from '../../getImgUrls';
