+++ src/templates/workspace/snippets/tree.ts
@@ -1,5 +1,5 @@
-import { getFileTooltip } from '../../../utils/strings/getFileTooltip';
 import { RenderVars } from '../../../webviews/webviews.interface';
+import { getFileTooltip } from '../../../webviews/Workspace/helpers/getFileTooltip';
 import { FileTree } from '../../../webviews/Workspace/helpers/getFileTree';
 import { File } from '../../../webviews/Workspace/WorkspaceViewProvider.interface';
 import { listItemButtons } from './listItemButtons';
